Get lambda function working, bring in terraform from ingestion-s3, updated makefile to create lambda layer
