+++ .gitignore
@@ -21,3 +21,4 @@ requirements.txt
 .secrets
 terraform/secret.tfvars
 requirements.txt
+layer/
