--- requirements.txt
@@ -1,117 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.12
-# by the following command:
-#
-#    pip-compile requirements.in
-#
-asn1crypto==1.5.1
-    # via scramp
-black==24.4.2
-    # via -r requirements.in
-boto3==1.34.105
-    # via
-    #   -r requirements.in
-    #   moto
-botocore==1.34.105
-    # via
-    #   boto3
-    #   moto
-    #   s3transfer
-certifi==2024.2.2
-    # via requests
-cffi==1.16.0
-    # via cryptography
-charset-normalizer==3.3.2
-    # via requests
-click==8.1.7
-    # via black
-coverage[toml]==7.5.1
-    # via
-    #   -r requirements.in
-    #   pytest-cov
-cryptography==42.0.7
-    # via moto
-flake8==7.0.0
-    # via -r requirements.in
-freezegun==1.5.1
-    # via -r requirements.in
-idna==3.7
-    # via requests
-iniconfig==2.0.0
-    # via pytest
-jinja2==3.1.4
-    # via moto
-jmespath==1.0.1
-    # via
-    #   boto3
-    #   botocore
-markupsafe==2.1.5
-    # via
-    #   jinja2
-    #   werkzeug
-mccabe==0.7.0
-    # via flake8
-mock==5.1.0
-    # via -r requirements.in
-moto==5.0.7
-    # via -r requirements.in
-mypy-extensions==1.0.0
-    # via black
-packaging==24.0
-    # via
-    #   black
-    #   pytest
-pathspec==0.12.1
-    # via black
-pg8000==1.31.2
-    # via -r requirements.in
-platformdirs==4.2.2
-    # via black
-pluggy==1.5.0
-    # via pytest
-pycodestyle==2.11.1
-    # via flake8
-pycparser==2.22
-    # via cffi
-pyflakes==3.2.0
-    # via flake8
-pytest==8.2.0
-    # via
-    #   -r requirements.in
-    #   pytest-cov
-    #   pytest-testdox
-pytest-cov==5.0.0
-    # via -r requirements.in
-pytest-testdox==3.1.0
-    # via -r requirements.in
-python-dateutil==2.9.0.post0
-    # via
-    #   botocore
-    #   freezegun
-    #   moto
-    #   pg8000
-python-dotenv==1.0.1
-    # via -r requirements.in
-pyyaml==6.0.1
-    # via responses
-requests==2.31.0
-    # via
-    #   moto
-    #   responses
-responses==0.25.0
-    # via moto
-s3transfer==0.10.1
-    # via boto3
-scramp==1.4.5
-    # via pg8000
-six==1.16.0
-    # via python-dateutil
-urllib3==2.2.1
-    # via
-    #   botocore
-    #   requests
-    #   responses
-werkzeug==3.0.3
-    # via moto
-xmltodict==0.13.0
-    # via moto
