+++ terraform/main.tf
@@ -12,8 +12,6 @@ terraform {
   }
 }
 
-
-
 provider "aws" {
   region = "eu-west-2"
   default_tags {
