+++ terraform/variables.tf
@@ -1,6 +1,7 @@
 variable "db_username" {
   description = "Database administrator username"
   type        = string
+  sensitive   = true
 }
  
 variable "db_password" {
